Archetypes for lecture 8

+++ spark-streaming-flights-archetype/src/main/resources/archetype-resources/src/main/scala/KafkaFlightRecord.scala
@@ -5,4 +5,4 @@ case class KafkaFlightRecord(
     flight: String,
     originName: String, 
     destinationName: String, 
-    departureDelay:Int)
+    departureDelay:Long)
